+++ merqube_client_lib/templates/equity_baskets/decrement/create.py
@@ -11,21 +11,11 @@ import pandas as pd
 
 from merqube_client_lib.exceptions import APIError
 from merqube_client_lib.logging import get_module_logger
+from merqube_client_lib.templates.equity_baskets.schema import ClientDecrementConfig
 from merqube_client_lib.templates.equity_baskets.util import create_index, load_template
 from merqube_client_lib.types import CreateReturn, Records
 
 logger = get_module_logger(__name__, level=logging.DEBUG)
-
-TYPE_SPECIFIC_REQUIRED = [
-    "underlying_index_name",
-    "fee_value",
-    "fee_type",
-    "day_count_convention",
-    "base_value",
-    "client_owned_underlying",
-]
-
-TYPE_SPECIFIC_OPTIONAL: list[str] = []
 
 
 def _download_tr_map(tdir: str) -> str:
@@ -63,14 +53,13 @@ def create(config_file_path: str, prod_run: bool = False) -> CreateReturn:
     This class does not handle Corax.
     """
     client, template, index_info, inner_spec = load_template(
-        template_name="DECREMENT_TEMPLATE_VERSION_1",
-        config_file_path=config_file_path,
-        type_specific_req_fields=TYPE_SPECIFIC_REQUIRED,
-        type_specific_opt_fields=TYPE_SPECIFIC_OPTIONAL,
+        template_name="DECREMENT_TEMPLATE_VERSION_1", config_file_path=config_file_path, model=ClientDecrementConfig
     )
 
-    uin = index_info["underlying_index_name"]
-    if index_info["client_owned_underlying"]:
+    index_info = cast(ClientDecrementConfig, index_info)
+
+    uin = index_info.underlying_index_name
+    if index_info.client_owned_underlying:
         # check that the underlying index exists
         logger.info(f"Checking that the underlying index {uin} exists")
         try:
@@ -89,19 +78,8 @@ def create(config_file_path: str, prod_run: bool = False) -> CreateReturn:
 
     inner_spec["metric"] = underlying_secapi_metric  # the decrement index looks up the returns of the tr via this
     inner_spec["underlying_ticker"] = uin
-
-    if (
-        not isinstance(fv := index_info["fee_value"], (int, float))
-        or not isinstance(ft := index_info["fee_type"], str)
-        or ft not in ["fixed", "percentage_pre", "percentage_post"]
-    ):
-        raise ValueError(
-            "fee_value must be a number and fee_type must be one of fixed, percentage_pre, percentage_post"
-        )
-
-    inner_spec["day_count_convention"] = index_info["day_count_convention"]
-
-    inner_spec["fee"] = {"fee_value": fv, "fee_type": ft}
+    inner_spec["day_count_convention"] = index_info.day_count_convention
+    inner_spec["fee"] = {"fee_value": index_info.fee_value, "fee_type": index_info.fee_type.value}
 
     post_template, ident_ppost = create_index(
         client=client, template=template, index_info=index_info, inner_spec=inner_spec, prod_run=prod_run